update tests structure, naming and such
